# doc/api-docs/openapi.yml
openapi: 3.0.3
info:
  title: myTrack API
  description: |-
    This document provides a detailed overview of the myTrack REST API.
    The API is used to manage IMEI-related resources, journeys, users, and other assets.
  version: 0.7.2-SNAPSHOT
servers:
  - url: https://thkrebs.ddns.net:8443/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development Server
...
paths:
...
  /user/{username}/features:
    get:
      tags:
        - User
      summary: Get User Features
      description: Retrieves the feature flags for a specific user. The feature field is a 64-bit integer that encodes the user's domain and subscribed packages.
      operationId: getUserFeatures
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The username of the user whose features are to be retrieved.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFeaturesDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
  /user/{username}/profile:
    get:
      tags:
        - User
      summary: Get User Profile
      description: Retrieves the full profile of a user. Accessible by the user themselves or an admin.
      operationId: getUserProfile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The username of the user whose profile is to be retrieved.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
    put:
      tags:
        - User
      summary: Update User Profile
      description: Updates the user's own profile (username and email only). Accessible only by the user themselves.
      operationId: updateUserProfile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The username of the user whose profile is to be updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateDTO'
      responses:
        '200':
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
  /user/{username}/admin/profile:
    put:
      tags:
        - User
      summary: Admin Update User Profile
      description: Updates the full user profile, including features and enabled status. Accessible only by admins (GOD role).
      operationId: adminUpdateUserProfile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: The username of the user whose profile is to be updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileAdminUpdateDTO'
      responses:
        '200':
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
  /user/resetPassword:
    post:
      tags:
        - User
      summary: Request Password Reset
      description: Initiates the password reset process by sending an email with a reset token to the user.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestDTO'
      responses:
        '200':
          description: Password reset email sent (if user exists).
          content:
            text/plain:
              schema:
                type: string
                example: "If an account with that email exists, a password reset link has been sent."
  /user/savePassword:
    post:
      tags:
        - User
      summary: Save New Password
      description: Updates the user's password using a valid reset token.
      operationId: savePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordSaveRequestDTO'
      responses:
        '200':
          description: Password updated successfully.
          content:
            text/plain:
              schema:
                type: string
                example: "Password reset successfully."
        '400':
          description: Invalid or expired token.

  # Journey Controller
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiTokenAuth: # Placeholder for API token security
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    UserFeaturesDTO:
      type: object
      properties:
        domain:
          type: integer
          format: int32
          description: The domain identifier for the user.
        packages:
          type: integer
          format: int32
          description: The package identifier for the user.
    UserProfileDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (yyyy-MM-dd HH:mm:ss)
        features:
          type: integer
          format: int64
        enabled:
          type: boolean
    UserProfileUpdateDTO:
      type: object
      required:
        - username
        - email
      properties:
        username:
          type: string
        email:
          type: string
          format: email
    UserProfileAdminUpdateDTO:
      type: object
      required:
        - username
        - email
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        features:
          type: integer
          format: int64
        enabled:
          type: boolean
...
    PasswordResetRequestDTO:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PasswordSaveRequestDTO:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string