# src/main/java/com/tmv/core/dto/UserProfileAdminUpdateDTO.java
package com.tmv.core.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class UserProfileAdminUpdateDTO {
    @NotBlank
    private String username;
    @NotBlank
    @Email
    private String email;
    private Long features;
    private boolean enabled;
}